requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
vaderSentiment==3.3.2
textblob==0.18.0
emoji==2.12.1
urlextract==1.9.0
werkzeug==3.0.3
scikit-learn==1.5.0
gunicorn==22.0.0
streamlit==1.35.0
plotly==5.22.0